Refactor fetchers
