+++ components/network/CircumventionToolsChart.js
@@ -1,9 +1,8 @@
 import React, { useMemo } from 'react'
 import { useRouter } from 'next/router'
 import { Heading, Box, Flex } from 'ooni-components'
-import axios from 'axios'
-import { axiosResponseTime } from 'components/axios-plugins'
 import useSWR from 'swr'
+import { MATMultipleFetcher as fetcher } from 'services/fetchers'
 import GridChart, { prepareDataForGridChart } from 'components/aggregation/mat/GridChart'
 import { DetailsBox } from 'components/measurement/DetailsBox'
 import { MATContextProvider } from 'components/aggregation/mat/MATContext'
@@ -12,39 +11,6 @@ import { FormattedMessage } from 'react-intl'
 const swrOptions = {
   revalidateOnFocus: false,
   dedupingInterval: 10 * 60 * 1000,
-}
-
-const baseURL = process.env.NEXT_PUBLIC_AGGREGATION_API
-axiosResponseTime(axios)
-
-const circumventionFetcher = (query) => {
-  const reqUrl = `${baseURL}/api/v1/aggregation?${query}`
-  const result = []
-
-  const request = (reqUrl) => { 
-    return axios.get(reqUrl).then(r => {
-      if (!r?.data?.result) {
-        const error = new Error(`Request ${reqUrl} did not contain expected result`)
-        error.data = r
-        throw error
-      }
-      return {
-        data: r.data.result,
-        loadTime: r.loadTime,
-        url: r.config.url
-      }
-    }).catch(e => {
-      e.message = e?.request?.response ?? e.message
-      throw e
-    })
-  }
-
-  const circumventionTools = ['psiphon', 'tor', 'torsf']
-  const requests = circumventionTools.map((tool) => request(`${reqUrl}&test_name=${tool}`))
-
-  return Promise.allSettled(requests).then((responses) => {
-    return responses.reduce((prev, current, index) => [...prev, ...current.value.data.map((obj) => ({...obj, circumvention_tool: circumventionTools[index]}))], [])
-  })
 }
 
 const testName = 'circumvention_tool'
@@ -58,7 +24,6 @@ const CircumventionChart = React.memo(function MessagingChart() {
     probe_asn: asn,
     since: since,
     until: until,
-    type: 'circumvention'
   }), [since, until, asn])
 
   const apiQuery = useMemo(() => {
@@ -67,22 +32,22 @@ const CircumventionChart = React.memo(function MessagingChart() {
   }, [query])
 
   const { data, error } = useSWR(
-    apiQuery,
-    circumventionFetcher,
+    { query: apiQuery, 
+      testNames: ['psiphon', 'tor', 'torsf'], 
+      groupKey: testName
+    },
+    fetcher,
     swrOptions
   )
+
   const [chartData, rowKeys, rowLabels] = useMemo(() => {
     if (!data) {
       return [null, 0]
     }
     let chartData = data
-
-    const graphQuery = {...query, axis_y: 'circumvention_tool'}
-
+    const graphQuery = {...query, axis_y: testName}
     const [reshapedData, rowKeys, rowLabels] = prepareDataForGridChart(chartData, graphQuery)
-
     return [reshapedData, rowKeys, rowLabels]
-
   }, [data, query])
 
   const headerOptions = { probe_cc: false, subtitle: false }
