+++ components/network/MessagingAppsChart.js
@@ -2,54 +2,17 @@ import React, { useMemo } from 'react'
 import { useRouter } from 'next/router'
 import { FormattedMessage } from 'react-intl'
 import { Heading, Box, Flex } from 'ooni-components'
-import axios from 'axios'
-
-import { axiosResponseTime } from 'components/axios-plugins'
 import useSWR from 'swr'
-
 import GridChart, { prepareDataForGridChart } from 'components/aggregation/mat/GridChart'
 import { DetailsBox } from 'components/measurement/DetailsBox'
 import { MATContextProvider } from 'components/aggregation/mat/MATContext'
+import { MATMultipleFetcher as fetcher } from 'services/fetchers'
 
 const swrOptions = {
   revalidateOnFocus: false,
   dedupingInterval: 10 * 60 * 1000,
 }
 
-const baseURL = process.env.NEXT_PUBLIC_AGGREGATION_API
-axiosResponseTime(axios)
-
-const messagingFetcher = (query) => {
-  const reqUrl = `${baseURL}/api/v1/aggregation?${query}`
-  const result = []
-
-  const request = (reqUrl) => { 
-    return axios.get(reqUrl).then(r => {
-      if (!r?.data?.result) {
-        const error = new Error(`Request ${reqUrl} did not contain expected result`)
-        error.data = r
-        throw error
-      }
-      return {
-        data: r.data.result,
-        loadTime: r.loadTime,
-        url: r.config.url
-      }
-    }).catch(e => {
-      console.log(e)
-      e.message = e?.request?.response ?? e.message
-      throw e
-    })
-  }
-
-  const messagingApps = ['signal', 'telegram', 'whatsapp', 'facebook_messenger']
-  const requests = messagingApps.map((tool) => request(`${reqUrl}&test_name=${tool}`))
-
-  return Promise.allSettled(requests).then((responses) => {
-    return responses.reduce((prev, current, index) => [...prev, ...current.value.data.map((obj) => ({...obj, messaging_app: messagingApps[index]}))], [])
-  })
-  
-}
 const testName = 'messaging_app'
 const queryParams = { axis_x: 'measurement_start_day' }
 
@@ -70,8 +33,11 @@ const MessagingChart = React.memo(function MessagingChart() {
   }, [query])
 
   const { data, error } = useSWR(
-    apiQuery,
-    messagingFetcher,
+    { query: apiQuery, 
+      testNames: ['signal', 'telegram', 'whatsapp', 'facebook_messenger'], 
+      groupKey: testName
+    },
+    fetcher,
     swrOptions
   )
   const [chartData, rowKeys, rowLabels] = useMemo(() => {
@@ -79,15 +45,10 @@ const MessagingChart = React.memo(function MessagingChart() {
       return [null, 0]
     }
     let chartData = data
-
-    const graphQuery = {...query, axis_y: 'messaging_app'}
-
+    const graphQuery = {...query, axis_y: testName}
     const [reshapedData, rowKeys, rowLabels] = prepareDataForGridChart(chartData, graphQuery)
-
     return [reshapedData, rowKeys, rowLabels]
-
   }, [data, query])
-
 
   const headerOptions = { probe_cc: false, subtitle: false }
