+++ components/network/WebConnectivityChart.js
@@ -1,40 +1,16 @@
 import React, { useCallback, useMemo } from 'react'
 import { useRouter } from 'next/router'
 import { Heading, Box, Flex } from 'ooni-components'
-import axios from 'axios'
-import { axiosResponseTime } from 'components/axios-plugins'
 import useSWR from 'swr'
 import GridChart, { prepareDataForGridChart } from 'components/aggregation/mat/GridChart'
 import { DetailsBox } from 'components/measurement/DetailsBox'
 import { MATContextProvider } from 'components/aggregation/mat/MATContext'
 import { FormattedMessage } from 'react-intl'
+import { MATFetcher as fetcher } from 'services/fetchers'
 
 const swrOptions = {
   revalidateOnFocus: false,
   dedupingInterval: 10 * 60 * 1000,
-}
-
-const baseURL = process.env.NEXT_PUBLIC_AGGREGATION_API
-axiosResponseTime(axios)
-
-const fetcher = (query) => {
-  const reqUrl = `${baseURL}/api/v1/aggregation?${query}`
-  return axios.get(reqUrl).then(r => {
-    if (!r?.data?.result) {
-      const error = new Error(`Request ${reqUrl} did not contain expected result`)
-      error.data = r
-      throw error
-    }
-    return {
-      data: r.data.result,
-      loadTime: r.loadTime,
-      url: r.config.url
-    }
-  }).catch(e => {
-    console.log(e)
-    e.message = e?.request?.response ?? e.message
-    throw e
-  })
 }
 
 const testName = 'web_connectivity'
@@ -43,9 +19,7 @@ const queryParams = { axis_x: 'measurement_start_day', axis_y: 'domain' }
 const Chart = React.memo(function Chart() {
   const router = useRouter()
   const { query: {since, until, asn} } = router
-  // Construct a `query` object that matches the router.query
-  // used by MAT, because in this case axis_x, axis_y are not part
-  // of router.query
+  
   const query = useMemo(() => ({
     ...queryParams,
     probe_asn: asn,
@@ -64,19 +38,15 @@ const Chart = React.memo(function Chart() {
     fetcher,
     swrOptions
   )
+
   const [chartData, rowKeys, rowLabels] = useMemo(() => {
     if (!data) {
       return [null, 0]
     }
-
     let chartData = data.data
-
     const [reshapedData, rowKeys, rowLabels] = prepareDataForGridChart(chartData, query)
-
     return [reshapedData, rowKeys, rowLabels]
-
   }, [data, query])
-
 
   const headerOptions = { probe_cc: false, subtitle: false }
