+++ package.json
@@ -50,7 +50,7 @@
     "react-window": "^1.8.6",
     "rebass": "4.0.7",
     "styled-components": "5.1.1",
-    "swr": "^0.2.3",
+    "swr": "1.3.0",
     "use-clipboard-copy": "^0.1.2",
     "victory": "^32.0.2"
   },
